Added cursorTracking as option with default as false
