+++ cypress.json
@@ -7,6 +7,7 @@
   "env": {
     "cypress-movie": {
       "enabled": true,
+      "cursorTracking": true,
       "width": 1920,
       "height": 1080
     }
