+++ src/cursor-tracking-commands/click.js
