+++ src/cursor-tracking-commands/default.js
