+++ src/cursor-tracking-commands/type.js
